add graphing to conditional
initially, an error would occur if no graphing argument was provided. this was remedied by adding the graphing situation to a conditional statement. this is done instead of prompting for graphing arguments in case the user does not want to graph (such as an instance that the user only wanted to cache their search"